fix: rm bugged unused Time and Absence benefits enrollment message
The code removed in this commit is bugged in that it's not predicated on the viewing employee actually having an enrollment opportunity, such that if it were not otherwise bugged as described below it would be incorrectly telling employees who do not in fact have an enrollment opportunity that they have an enrollment opportunity.

Fortunately, the code removed in this commit is also bugged in that it uses and is predicated on the presence of an HRS URL `Benefits Enrollment` but that's not the key that HRS uses to convey the URL for self-service benefits enrollment, so the c:if test case here never evaluates true and so this div is never displayed.

Hurray for saving graces. This commit therefore amounts to removing distracting no-op code.