fix: inline generated markup rather than rely on AngularJS directives
AngularJS directives don't appear to be working in remote-content content.
It'd be a good idea to get that working, or even better to get use of Web Component successors to the directives working, so that consistent UI components can be used in front-end and back-end sourced widgets. This kludge should get the widget working for immediate purposes.