fix: aria-label the "Learn more" links
"Learn more" links fall into the "click here" link anti-pattern: when jumping to the link out of context, it will be unclear from just the link text "Learn more" what the link would have one learn more about. The aria-label fills in the context implied by the visual presentation of the link in the context of Benefit Information and enrollment.